requirement.txt
absl-py==0.7.0
astor==0.7.1
beautifulsoup4==4.7.1
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
grpcio==1.18.0
h5py==2.9.0
idna==2.8
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==3.0.2
mock==2.0.0
numpy==1.16.1
pbr==5.1.2
Pillow==5.4.1
protobuf==3.6.1
pyparsing==2.3.1
pyspark==2.1.1+hadoop2.7
python-dateutil==2.8.0
PyYAML==3.13
requests==2.21.0
scipy==1.2.1
selenium==3.141.0
six==1.12.0
soupsieve==1.8
TaiwanTrainVerificationCode2text==1.0.16
tensorboard==1.12.2
tensorflow==1.13.0rc2
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.24.1
Werkzeug==0.14.1
